start abstracting requirement scanner
